Completed task 6

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long long int c;
 	float d;
 
-	printf("Size of a char: %d byte(s)\n", (int)sizeof(a));
-	printf("Size of an int: %d byte(s)\n", (int)sizeof(i));
-	printf("Size of a long int: %d byte(s)\n", (int)sizeof(b));
-	printf("Size of a long long int: %d bytes(s)\n", (int)sizeof(c));
-	printf("Size of a float: %d byte(s)\n", (int)sizeof(d));
+	printf("Size of a char: %lu byte(s)\n", sizeof(a));
+	printf("Size of an int: %lu byte(s)\n", sizeof(i));
+	printf("Size of a long int: %lu byte(s)\n", sizeof(b));
+	printf("Size of a long long int: %lu bytes(s)\n", sizeof(c));
+	printf("Size of a float: %lu byte(s)\n",sizeof(d));
 	return (0);
 }
